Poprawienie bledow w projekcie i pseudokodzie

+++ Inne projekty/Zamiana dwoch liczb/metody/swap8.c
@@ -2,19 +2,19 @@
 int main (void)
 {
 
-int a,b, c, d;
+float a,b, c, d;
 int i;
 for (i = 1; i <=10; i++)
 {
-	scanf("%d %d", &a, &b);
+	scanf("%g %g", &a, &b);
 	c = a;
 	d = b;
 	
 	c = c / d;
-	d = c * d;
+	d = c * d; 
 	c = d / c;
 	
-	printf("%d %d\n", c, d);
+	printf("%g %g\n", c, d);
 	
 }
 return 0;
